Journey Journal Application
This was an project that I successfully executed earlier in the year of an android app development. The application was developed to open a user account with the authentication of an email address and a password which takes the user to the index page after a successful sign up.  The database for the application was developed with SQLite and java was the language used for the functionality of the android application. The application has the capacity of indicating and recording the geopolitical location of the user. The user can record his experience while on the journey, just as one takes record in a diary and then sign out after experience.